+++ BraintreeCore/src/test/java/com/braintreepayments/api/BraintreeClientUnitTest.kt
@@ -635,47 +635,11 @@ class BraintreeClientUnitTest {
     }
 
     @Test
-    fun sessionId_withAuthString_returnsSessionIdDefinedInConstructor() {
-        val context = ApplicationProvider.getApplicationContext<Context>()
-        val authorization = Fixtures.BASE64_CLIENT_TOKEN
-        val sessionId = "custom-session-id"
-        val sut = BraintreeClient(context, authorization, sessionId, IntegrationType.DROP_IN)
-        assertEquals("custom-session-id", sut.sessionId)
-    }
-
-    @Test
-    fun sessionId_withClientTokenProvider_returnsSessionIdDefinedInConstructor() {
-        val context = ApplicationProvider.getApplicationContext<Context>()
-        val clientTokenProvider = mockk<ClientTokenProvider>(relaxed = true)
-        val sessionId = "custom-session-id"
-        val sut = BraintreeClient(context, clientTokenProvider, sessionId, IntegrationType.DROP_IN)
-        assertEquals("custom-session-id", sut.sessionId)
-    }
-
-    @Test
     fun integrationType_returnsCustomByDefault() {
         val context = ApplicationProvider.getApplicationContext<Context>()
         val authorization = Fixtures.BASE64_CLIENT_TOKEN
         val sut = BraintreeClient(context, authorization)
         assertEquals("custom", sut.integrationType)
-    }
-
-    @Test
-    fun integrationType_withAuthString_returnsIntegrationTypeDefinedInConstructor() {
-        val context = ApplicationProvider.getApplicationContext<Context>()
-        val authorization = Fixtures.BASE64_CLIENT_TOKEN
-        val sessionId = "custom-session-id"
-        val sut = BraintreeClient(context, authorization, sessionId, IntegrationType.DROP_IN)
-        assertEquals("dropin", sut.integrationType)
-    }
-
-    @Test
-    fun integrationType_withClientTokenProvider_returnsIntegrationTypeDefinedInConstructor() {
-        val context = ApplicationProvider.getApplicationContext<Context>()
-        val clientTokenProvider = mockk<ClientTokenProvider>(relaxed = true)
-        val sessionId = "custom-session-id"
-        val sut = BraintreeClient(context, clientTokenProvider, sessionId, IntegrationType.DROP_IN)
-        assertEquals("dropin", sut.integrationType)
     }
 
     @Test
@@ -715,6 +679,48 @@ class BraintreeClientUnitTest {
         }
     }
 
+    @Test
+    fun `when client is created with authorization, app link is set on appLinkReturnUri`() {
+        val appLinkUrl = "https://merchant-site.com"
+        val sut: BraintreeClient = BraintreeClient(
+            context = context,
+            authorization = "authorization",
+            returnUrlScheme = "returnUrlScheme",
+            appLinkReturnUri = Uri.parse(appLinkUrl)
+        )
+        assertEquals(appLinkUrl, sut.appLinkReturnUri.toString())
+    }
+
+    @Test
+    fun `when client is created with ClientTokenProvider, app link is set on appLinkReturnUri`() {
+        val appLinkUrl = "https://merchant-site.com"
+        val sut: BraintreeClient = BraintreeClient(
+            context = context,
+            clientTokenProvider = mockk(),
+            returnUrlScheme = "returnUrlScheme",
+            appLinkReturnUri = Uri.parse(appLinkUrl)
+        )
+        assertEquals(appLinkUrl, sut.appLinkReturnUri.toString())
+    }
+
+    @Test
+    fun `when client is created with authorization without an app link, appLinkReturnUri is null`() {
+        val sut: BraintreeClient = BraintreeClient(
+            context = context,
+            authorization = "authorization"
+        )
+        assertNull(sut.appLinkReturnUri)
+    }
+
+    @Test
+    fun `when client is created with ClientTokenProvider without an app link, appLinkReturnUri is null`() {
+        val sut: BraintreeClient = BraintreeClient(
+            context = context,
+            clientTokenProvider = mockk()
+        )
+        assertNull(sut.appLinkReturnUri)
+    }
+
     private fun createDefaultParams(
         configurationLoader: ConfigurationLoader,
         authorizationLoader: AuthorizationLoader
@@ -730,7 +736,8 @@ class BraintreeClientUnitTest {
             browserSwitchClient = browserSwitchClient,
             manifestValidator = manifestValidator,
             configurationLoader = configurationLoader,
-            integrationType = IntegrationType.CUSTOM
+            integrationType = IntegrationType.CUSTOM,
+            appLinkReturnUri = Uri.parse("https://sample-merchant-site.com"),
         )
 
     companion object {
@@ -738,8 +745,9 @@ class BraintreeClientUnitTest {
             val requestCode = 123
             val url = Uri.parse("www.example.com")
             val returnUrlScheme = "sample-scheme"
+            val appLinkUri = Uri.parse("sample-scheme")
             val browserSwitchRequest = BrowserSwitchRequest(
-                requestCode, url, JSONObject(), returnUrlScheme, true
+                requestCode, url, JSONObject(), returnUrlScheme, appLinkUri, true
             )
             return BrowserSwitchResult(BrowserSwitchStatus.SUCCESS, browserSwitchRequest)
         }