+++ PayPal/src/test/java/com/braintreepayments/api/PayPalCheckoutRequestUnitTest.java
@@ -34,6 +34,7 @@ public class PayPalCheckoutRequestUnitTest {
         assertNull(request.getBillingAgreementDescription());
         assertFalse(request.getShouldOfferPayLater());
         assertFalse(request.hasUserLocationConsent());
+        assertFalse(request.isAppLinkEnabled());
     }
 
     @Test
@@ -59,6 +60,7 @@ public class PayPalCheckoutRequestUnitTest {
         request.setDisplayName("Display Name");
         request.setRiskCorrelationId("123-correlation");
         request.setLandingPageType(PayPalRequest.LANDING_PAGE_TYPE_LOGIN);
+        request.setAppLinkEnabled(true);
 
         assertEquals("1.00", request.getAmount());
         assertEquals("USD", request.getCurrencyCode());
@@ -74,6 +76,7 @@ public class PayPalCheckoutRequestUnitTest {
         assertEquals(PayPalRequest.LANDING_PAGE_TYPE_LOGIN, request.getLandingPageType());
         assertTrue(request.getShouldOfferPayLater());
         assertTrue(request.hasUserLocationConsent());
+        assertTrue(request.isAppLinkEnabled());
     }
 
     @Test
@@ -95,6 +98,7 @@ public class PayPalCheckoutRequestUnitTest {
         request.setDisplayName("Display Name");
         request.setRiskCorrelationId("123-correlation");
         request.setMerchantAccountId("merchant_account_id");
+        request.setAppLinkEnabled(true);
 
         ArrayList<PayPalLineItem> lineItems = new ArrayList<>();
         lineItems.add(new PayPalLineItem(PayPalLineItem.KIND_DEBIT, "An Item", "1", "1"));
@@ -123,6 +127,7 @@ public class PayPalCheckoutRequestUnitTest {
         assertEquals(1, result.getLineItems().size());
         assertEquals("An Item", result.getLineItems().get(0).getName());
         assertTrue(result.hasUserLocationConsent());
+        assertTrue(result.isAppLinkEnabled());
     }
 
     @Test